+++ Dockerfile
@@ -1,5 +1,10 @@
-FROM tiangolo/uwsgi-nginx-flask:python3.8
-
+FROM python:buster
 COPY . /app
+WORKDIR /app
+RUN python -m venv venv 
+RUN venv/bin/pip install -r requirements.txt 
+RUN venv/bin/pip install gunicorn
+RUN chmod +x start.sh
+EXPOSE 5000
+CMD ["./start.sh"]
 
-RUN pip install -r requirements.txt
