+++ app/insults.txt
