+++ app/service.py
@@ -1,0 +1,21 @@
+from app.flippyflop import FlippyFlop
+import pickle
+from googleapiclient.discovery import build
+import datetime
+
+credential_path = "app/auth/token.pickle"
+with open(credential_path, "rb") as token:
+    creds = pickle.load(token)
+
+service = build("sheets", "v4", credentials=creds)
+spreadsheet_id = "1eZL2eOCFKxGkg7bYaEmp-urWqWBfUNx73n_1oR2RkpM"
+
+ff = FlippyFlop(
+    service=service,
+    spreadsheet_id=spreadsheet_id,
+    day_zero=datetime.datetime(2020, 7, 7),
+    throttle_time=0,
+)
+
+REMAINING = ff.todays_cards()
+TERMS = ff.get_terms()
